heat-map/requirements.txt
quixstreams==2.6.0
python-dotenv